fix; norm ok in builtin
