+++ srcs/builtin/export_utils_2.c
@@ -1,33 +1,49 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   env.c                                              :+:      :+:    :+:   */
+/*   export_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hnakayam <hnakayam@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/11/24 15:26:01 by hnakayam          #+#    #+#             */
-/*   Updated: 2024/11/24 15:26:02 by hnakayam         ###   ########.fr       */
+/*   Created: 2024/11/24 16:22:57 by hnakayam          #+#    #+#             */
+/*   Updated: 2024/11/24 16:22:58 by hnakayam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 #include "builtin.h"
+#include <ctype.h>
 
-int	builtin_env(t_env **env, char **args)
+int	is_alpha_under(char c)
 {
-	t_env	*top;
-
-	top = *env;
-	if (args[0] != NULL)
-	{
-		ft_printf_stderr("Syntax Error\n");
+	if (isalpha(c) || c == '_')
 		return (1);
-	}
-	while (top)
-	{
-		if (top->value)
-			ft_printf("%s=%s\n", top->key, top->value);
-		top = top->next;
-	}
 	return (0);
+}
+
+int	is_alpha_num_under(char c)
+{
+	if (isalpha(c) || isdigit(c) || c == '_')
+		return (1);
+	return (0);
+}
+
+int	is_valid_key(char *key)
+{
+	if (!is_alpha_under(*key))
+		return (0);
+	key++;
+	while (*key)
+	{
+		if (!is_alpha_num_under(*key))
+			return (0);
+		key++;
+	}
+	return (1);
+}
+
+void	malloc_error_exit(void)
+{
+	write(2, "Error: malloc\n", 14);
+	exit(1);
 }
